+++ src/app/api/mcp/route.ts
@@ -2,6 +2,11 @@ import {WebStandardStreamableHTTPServerTransport} from '@modelcontextprotocol/sd
 
 import createMcpServer from '@/mcp/server';
 
+/**
+ * @route POST /api/mcp
+ * @route DELETE /api/mcp
+ * @description MCP Streamable HTTP 엔드포인트 — stateless (세션 없음)
+ */
 const handleRequest = async (req: Request) => {
   const server = createMcpServer();
   const transport = new WebStandardStreamableHTTPServerTransport({
