+++ src/app/api/picks/route.ts
@@ -5,12 +5,12 @@ import {PickResult, PickType} from '@/api/vote.types';
 
 const PICK_TYPES: PickType[] = ['COURSE_1', 'COURSE_2', 'TAKE_OUT', 'pass'];
 
-const emptyResult = (date: string, myPick: PickType | null = null): PickResult => ({
-  date,
-  counts: {COURSE_1: 0, COURSE_2: 0, TAKE_OUT: 0, pass: 0},
-  myPick,
-});
-
+/**
+ * @route GET /api/picks
+ * @header x-voter-id - 익명 투표자 ID
+ * @query date - 조회할 날짜 (YYYY-MM-DD)
+ * @returns 해당 날짜의 코스 픽 집계 및 내 선택
+ */
 export const GET = async (req: NextRequest) => {
   const date = req.nextUrl.searchParams.get('date');
   if (!date) {
@@ -19,13 +19,15 @@ export const GET = async (req: NextRequest) => {
 
   const voterId = req.headers.get('x-voter-id') ?? '';
 
+  const empty: PickResult = {date, counts: {COURSE_1: 0, COURSE_2: 0, TAKE_OUT: 0, pass: 0}, myPick: null};
+
   try {
     const {data, error} = await supabaseServer
       .from('menu_picks')
       .select('pick_type, voter_id')
       .eq('date', date);
 
-    if (error) return NextResponse.json(emptyResult(date), {status: 200});
+    if (error) return NextResponse.json(empty, {status: 200});
 
     const counts: Record<PickType, number> = {COURSE_1: 0, COURSE_2: 0, TAKE_OUT: 0, pass: 0};
     let myPick: PickType | null = null;
@@ -38,10 +40,16 @@ export const GET = async (req: NextRequest) => {
 
     return NextResponse.json({date, counts, myPick} satisfies PickResult);
   } catch {
-    return NextResponse.json(emptyResult(date), {status: 200});
+    return NextResponse.json(empty, {status: 200});
   }
 };
 
+/**
+ * @route POST /api/picks
+ * @header x-voter-id - 익명 투표자 ID
+ * @body `{ date, pick_type }` — pick_type이 null이면 기존 픽 취소
+ * @returns `{ ok: true }`
+ */
 export const POST = async (req: NextRequest) => {
   const voterId = req.headers.get('x-voter-id') ?? '';
   if (!voterId) {
