+++ src/app/api/slack/warm/route.ts
@@ -3,6 +3,11 @@ import {NextRequest, NextResponse} from 'next/server';
 
 import {WARM_KEYWORDS} from '@/constants/slack';
 
+/**
+ * @route GET /api/slack/warm
+ * @description Vercel Cron이 매일 15:00 UTC 호출 — '/' 경로와 'menu' 태그 재검증 후 Slack 응답 캐시 워밍
+ * @returns `{ results }` — 키워드별 워밍 결과 목록
+ */
 export async function GET(req: NextRequest) {
   revalidatePath('/');
   revalidateTag('menu');
